requirements.txt
requests>=2.31.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
